move extension to test file
