+++ android/src/test/java/com/reactnativecommunity/asyncstorage/next/ArgumentHelpersTest.kt
@@ -1,6 +1,7 @@
 package com.reactnativecommunity.asyncstorage.next
 
 import com.facebook.react.bridge.JavaOnlyArray
+import com.facebook.react.bridge.ReadableArray
 import com.google.common.truth.Truth.assertThat
 import org.junit.Assert.assertThrows
 import org.junit.Test
@@ -76,5 +77,17 @@ class ArgumentHelpersTest {
     }
 }
 
-
-
+fun List<Any?>.toReadableArray(): ReadableArray {
+    val arr = JavaOnlyArray()
+    forEach {
+        when (it) {
+            null -> arr.pushNull()
+            is Boolean -> arr.pushBoolean(it)
+            is Double -> arr.pushDouble(it)
+            is Int -> arr.pushInt(it)
+            is String -> arr.pushString(it)
+            else -> throw NotImplementedError()
+        }
+    }
+    return arr
+}
