--- android/src/test/java/com/reactnativecommunity/asyncstorage/next/TestUtils.kt
@@ -1,19 +1,0 @@
-package com.reactnativecommunity.asyncstorage.next
-
-import com.facebook.react.bridge.JavaOnlyArray
-import com.facebook.react.bridge.ReadableArray
-
-fun List<Any?>.toReadableArray(): ReadableArray {
-    val arr = JavaOnlyArray()
-    forEach {
-        when (it) {
-            null -> arr.pushNull()
-            is Boolean -> arr.pushBoolean(it)
-            is Double -> arr.pushDouble(it)
-            is Int -> arr.pushInt(it)
-            is String -> arr.pushString(it)
-            else -> throw NotImplementedError()
-        }
-    }
-    return arr
-}
